+++ hadoop-tools/hadoop-distcp/src/main/java/org/apache/hadoop/tools/util/DistCpUtils.java
@@ -18,12 +18,20 @@
 
 package org.apache.hadoop.tools.util;
 
-import com.google.common.collect.Maps;
+import java.io.IOException;
+import java.net.InetAddress;
+import java.net.URI;
+import java.net.UnknownHostException;
+import java.text.DecimalFormat;
+import java.util.EnumSet;
+import java.util.List;
+import java.util.Locale;
+import java.util.Map;
+import java.util.Map.Entry;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.BlockLocation;
 import org.apache.hadoop.fs.FileChecksum;
 import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
@@ -32,25 +40,18 @@ import org.apache.hadoop.fs.XAttr;
 import org.apache.hadoop.fs.permission.AclEntry;
 import org.apache.hadoop.fs.permission.AclUtil;
 import org.apache.hadoop.fs.permission.FsPermission;
-import org.apache.hadoop.hdfs.DistributedFileSystem;
 import org.apache.hadoop.io.SequenceFile;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.InputFormat;
 import org.apache.hadoop.tools.CopyListing.AclsNotSupportedException;
 import org.apache.hadoop.tools.CopyListing.XAttrsNotSupportedException;
 import org.apache.hadoop.tools.CopyListingFileStatus;
-import org.apache.hadoop.tools.DistCpContext;
+import org.apache.hadoop.tools.DistCpOptions;
 import org.apache.hadoop.tools.DistCpOptions.FileAttribute;
 import org.apache.hadoop.tools.mapred.UniformSizeInputFormat;
-import org.apache.hadoop.util.StringUtils;
 
-import java.io.IOException;
-import java.text.DecimalFormat;
-import java.util.EnumSet;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
+import com.google.common.collect.Maps;
+import org.apache.hadoop.util.StringUtils;
 
 /**
  * Utility functions used in DistCp.
@@ -116,13 +117,13 @@ public class DistCpUtils {
    * a particular strategy from distcp-default.xml
    *
    * @param conf - Configuration object
-   * @param context - Distcp context with associated input options
+   * @param options - Handle to input options
    * @return Class implementing the strategy specified in options.
    */
   public static Class<? extends InputFormat> getStrategy(Configuration conf,
-      DistCpContext context) {
+                                                                 DistCpOptions options) {
     String confLabel = "distcp."
-        + StringUtils.toLowerCase(context.getCopyStrategy())
+        + StringUtils.toLowerCase(options.getCopyStrategy())
         + ".strategy" + ".impl";
     return conf.getClass(confLabel, UniformSizeInputFormat.class, InputFormat.class);
   }
@@ -199,13 +200,9 @@ public class DistCpUtils {
                               EnumSet<FileAttribute> attributes,
                               boolean preserveRawXattrs) throws IOException {
 
-    // If not preserving anything from FileStatus, don't bother fetching it.
-    FileStatus targetFileStatus = attributes.isEmpty() ? null :
-        targetFS.getFileStatus(path);
-    String group = targetFileStatus == null ? null :
-        targetFileStatus.getGroup();
-    String user = targetFileStatus == null ? null :
-        targetFileStatus.getOwner();
+    FileStatus targetFileStatus = targetFS.getFileStatus(path);
+    String group = targetFileStatus.getGroup();
+    String user = targetFileStatus.getOwner();
     boolean chown = false;
 
     if (attributes.contains(FileAttribute.ACL)) {
@@ -240,13 +237,8 @@ public class DistCpUtils {
       }
     }
 
-    // The replication factor can only be preserved for replicated files.
-    // It is ignored when either the source or target file are erasure coded.
-    if (attributes.contains(FileAttribute.REPLICATION) &&
-        !targetFileStatus.isDirectory() &&
-        !targetFileStatus.isErasureCoded() &&
-        !srcFileStatus.isErasureCoded() &&
-        srcFileStatus.getReplication() != targetFileStatus.getReplication()) {
+    if (attributes.contains(FileAttribute.REPLICATION) && !targetFileStatus.isDirectory() &&
+        (srcFileStatus.getReplication() != targetFileStatus.getReplication())) {
       targetFS.setReplication(path, srcFileStatus.getReplication());
     }
 
@@ -302,86 +294,6 @@ public class DistCpUtils {
   }
 
   /**
-   * Converts FileStatus to a list of CopyListingFileStatus.
-   * The resulted list contains either one CopyListingFileStatus per chunk of
-   * file-blocks (if file-size exceeds blockSize * blocksPerChunk, and there
-   * are more blocks in the file than blocksperChunk), or a single
-   * CopyListingFileStatus for the entire file (if file-size is too small to
-   * split).
-   * If preserving ACLs, populates the CopyListingFileStatus with the ACLs.
-   * If preserving XAttrs, populates the CopyListingFileStatus with the XAttrs.
-   *
-   * @param fileSystem FileSystem containing the file
-   * @param fileStatus FileStatus of file
-   * @param preserveAcls boolean true if preserving ACLs
-   * @param preserveXAttrs boolean true if preserving XAttrs
-   * @param preserveRawXAttrs boolean true if preserving raw.* XAttrs
-   * @param blocksPerChunk size of chunks when copying chunks in parallel
-   * @return list of CopyListingFileStatus
-   * @throws IOException if there is an I/O error
-   */
-  public static LinkedList<CopyListingFileStatus> toCopyListingFileStatus(
-      FileSystem fileSystem, FileStatus fileStatus, boolean preserveAcls,
-      boolean preserveXAttrs, boolean preserveRawXAttrs, int blocksPerChunk)
-          throws IOException {
-    LinkedList<CopyListingFileStatus> copyListingFileStatus =
-        new LinkedList<CopyListingFileStatus>();
-
-    final CopyListingFileStatus clfs = toCopyListingFileStatusHelper(
-        fileSystem, fileStatus, preserveAcls,
-        preserveXAttrs, preserveRawXAttrs,
-        0, fileStatus.getLen());
-    final long blockSize = fileStatus.getBlockSize();
-    if (LOG.isDebugEnabled()) {
-      LOG.debug("toCopyListing: " + fileStatus + " chunkSize: "
-          + blocksPerChunk + " isDFS: " +
-          (fileSystem instanceof DistributedFileSystem));
-    }
-    if ((blocksPerChunk > 0) &&
-        !fileStatus.isDirectory() &&
-        (fileStatus.getLen() > blockSize * blocksPerChunk)) {
-      // split only when the file size is larger than the intended chunk size
-      final BlockLocation[] blockLocations;
-      blockLocations = fileSystem.getFileBlockLocations(fileStatus, 0,
-            fileStatus.getLen());
-
-      int numBlocks = blockLocations.length;
-      long curPos = 0;
-      if (numBlocks <= blocksPerChunk) {
-        if (LOG.isDebugEnabled()) {
-          LOG.debug("  add file " + clfs);
-        }
-        copyListingFileStatus.add(clfs);
-      } else {
-        int i = 0;
-        while (i < numBlocks) {
-          long curLength = 0;
-          for (int j = 0; j < blocksPerChunk && i < numBlocks; ++j, ++i) {
-            curLength += blockLocations[i].getLength();
-          }
-          if (curLength > 0) {
-            CopyListingFileStatus clfs1 = new CopyListingFileStatus(clfs);
-            clfs1.setChunkOffset(curPos);
-            clfs1.setChunkLength(curLength);
-            if (LOG.isDebugEnabled()) {
-              LOG.debug("  add file chunk " + clfs1);
-            }
-            copyListingFileStatus.add(clfs1);
-            curPos += curLength;
-          }
-        }
-      }
-    } else {
-      if (LOG.isDebugEnabled()) {
-        LOG.debug("  add file/dir " + clfs);
-      }
-      copyListingFileStatus.add(clfs);
-    }
-
-    return copyListingFileStatus;
-  }
-
-  /**
    * Converts a FileStatus to a CopyListingFileStatus.  If preserving ACLs,
    * populates the CopyListingFileStatus with the ACLs. If preserving XAttrs,
    * populates the CopyListingFileStatus with the XAttrs.
@@ -391,17 +303,13 @@ public class DistCpUtils {
    * @param preserveAcls boolean true if preserving ACLs
    * @param preserveXAttrs boolean true if preserving XAttrs
    * @param preserveRawXAttrs boolean true if preserving raw.* XAttrs
-   * @param chunkOffset chunk offset in bytes
-   * @param chunkLength chunk length in bytes
-   * @return CopyListingFileStatus
    * @throws IOException if there is an I/O error
    */
-  public static CopyListingFileStatus toCopyListingFileStatusHelper(
+  public static CopyListingFileStatus toCopyListingFileStatus(
       FileSystem fileSystem, FileStatus fileStatus, boolean preserveAcls, 
-      boolean preserveXAttrs, boolean preserveRawXAttrs,
-      long chunkOffset, long chunkLength) throws IOException {
+      boolean preserveXAttrs, boolean preserveRawXAttrs) throws IOException {
     CopyListingFileStatus copyListingFileStatus =
-        new CopyListingFileStatus(fileStatus, chunkOffset, chunkLength);
+      new CopyListingFileStatus(fileStatus);
     if (preserveAcls) {
       FsPermission perm = fileStatus.getPermission();
       if (perm.getAclBit()) {
@@ -449,7 +357,9 @@ public class DistCpUtils {
       CopyListingFileStatus.class, conf);
     Path output = new Path(sourceListing.toString() +  "_sorted");
 
-    fs.delete(output, false);
+    if (fs.exists(output)) {
+      fs.delete(output, false);
+    }
 
     sorter.sort(sourceListing, output);
     return output;
@@ -496,7 +406,7 @@ public class DistCpUtils {
   /**
    * String utility to convert a number-of-bytes to human readable format.
    */
-  private static final ThreadLocal<DecimalFormat> FORMATTER
+  private static ThreadLocal<DecimalFormat> FORMATTER
                         = new ThreadLocal<DecimalFormat>() {
     @Override
     protected DecimalFormat initialValue() {
@@ -559,18 +469,42 @@ public class DistCpUtils {
             sourceChecksum.equals(targetChecksum));
   }
 
-  /*
-   * Return the Path for a given chunk.
-   * Used when splitting large file into chunks to copy in parallel.
-   * @param targetFile path to target file
-   * @param srcFileStatus source file status in copy listing
-   * @return path to the chunk specified by the parameters to store
-   * in target cluster temporarily
+  /* see if two file systems are the same or not
+   *
    */
-  public static Path getSplitChunkPath(Path targetFile,
-      CopyListingFileStatus srcFileStatus) {
-    return new Path(targetFile.toString()
-        + ".____distcpSplit____" + srcFileStatus.getChunkOffset()
-        + "." + srcFileStatus.getChunkLength());
+  public static boolean compareFs(FileSystem srcFs, FileSystem destFs) {
+    URI srcUri = srcFs.getUri();
+    URI dstUri = destFs.getUri();
+    if (srcUri.getScheme() == null) {
+      return false;
+    }
+    if (!srcUri.getScheme().equals(dstUri.getScheme())) {
+      return false;
+    }
+    String srcHost = srcUri.getHost();
+    String dstHost = dstUri.getHost();
+    if ((srcHost != null) && (dstHost != null)) {
+      try {
+        srcHost = InetAddress.getByName(srcHost).getCanonicalHostName();
+        dstHost = InetAddress.getByName(dstHost).getCanonicalHostName();
+      } catch(UnknownHostException ue) {
+        if (LOG.isDebugEnabled())
+          LOG.debug("Could not compare file-systems. Unknown host: ", ue);
+        return false;
+      }
+      if (!srcHost.equals(dstHost)) {
+        return false;
+      }
+    }
+    else if (srcHost == null && dstHost != null) {
+      return false;
+    }
+    else if (srcHost != null) {
+      return false;
+    }
+
+    //check for ports
+
+    return srcUri.getPort() == dstUri.getPort();
   }
 }
